Squashing similar tests
