+++ tests/cli_tests/zwalletcli_zcnbridge_burn_test.go
@@ -1,0 +1,47 @@
+package cli_tests
+
+import (
+	"fmt"
+	"strings"
+	"testing"
+
+	cliutils "github.com/0chain/system_test/internal/cli/util"
+	"github.com/stretchr/testify/require"
+)
+
+func TestBridgeBurn(t *testing.T) {
+	t.Parallel()
+
+	var zwallet = func(cmd, amount, help string) ([]string, error) {
+		run := fmt.Sprintf("./zwallet %s --amount %s", cmd, amount)
+		t.Logf("%s: amount %s", help, amount)
+		t.Log(run)
+		return cliutils.RunCommandWithoutRetry(run)
+	}
+
+	t.Run("Burning ZCN tokens", func(t *testing.T) {
+		t.Parallel()
+
+		output, err := zwallet(
+			"bridge-burn-zcn",
+			"1",
+			"Burning ZCN tokens that will be minted for WZCN tokens",
+		)
+
+		require.Nil(t, err, "error trying to burn ZCN tokens: %s", strings.Join(output, "\n"))
+		require.Contains(t, output[len(output)-1], "Verification successful")
+	})
+
+	t.Run("Burning WZCN tokens", func(t *testing.T) {
+		t.Parallel()
+
+		output, err := zwallet(
+			"bridge-burn-eth",
+			"10",
+			"Burning WZCN tokens that will be minted for ZCN tokens",
+		)
+
+		require.Nil(t, err, "error trying to burn WZCN tokens: %s", strings.Join(output, "\n"))
+		require.Contains(t, output[len(output)-1], "Verification: WZCN burn [OK]")
+	})
+}
